requirements.txt
PyMySQL==1.0.2
python-dotenv==0.19.0
SQLAlchemy==1.4.23
pandas
